+++ pkg/infra/main_helper.go
@@ -171,11 +171,6 @@ func updateInternalStates(immutableFightRounds *commons.ImmutableList[decision.I
 	Hp Pool Helpers
 */
 
-type HpPoolDonation struct {
-	agentId  commons.ID
-	donation uint
-}
-
 func checkHpPool() {
 	if globalState.HpPool >= globalState.MonsterAttack {
 		logging.Log(logging.Info, logging.LogField{
@@ -187,53 +182,4 @@ func checkHpPool() {
 		globalState.HpPool -= globalState.MonsterAttack
 		globalState.MonsterHealth = 0
 	}
-}
-
-func updateHpPool() {
-	var wg sync.WaitGroup
-	donationChan := make(chan HpPoolDonation, len(agentMap))
-	for id, a := range agentMap {
-		// fmt.Print(a.BaseAgent.AgentState().Hp)
-		id := id
-		a := a
-		wg.Add(1)
-		go func(wait *sync.WaitGroup, donationChan chan HpPoolDonation) {
-			donation := a.HandleDonateToHpPool(globalState.AgentState[id])
-			donationChan <- HpPoolDonation{id, donation}
-			wait.Done()
-		}(&wg, donationChan)
-	}
-	wg.Wait()
-
-	sum := uint(0)
-	for i := 0; i < len(agentMap); i++ {
-		agentDonation := <-donationChan
-
-		agentHp := globalState.AgentState[agentDonation.agentId].Hp
-		if agentDonation.donation >= agentHp {
-			agentDonation.donation = agentHp
-			delete(globalState.AgentState, agentDonation.agentId)
-			delete(agentMap, agentDonation.agentId)
-		}
-
-		logging.Log(logging.Trace, logging.LogField{
-			"Agent Donation": agentDonation,
-			"Old Sum":        sum,
-			"New Sum":        sum + agentDonation.donation,
-		}, "HP Pool Donation")
-
-		sum += agentDonation.donation
-		if a, ok := globalState.AgentState[agentDonation.agentId]; ok {
-			a.Hp = agentHp - agentDonation.donation
-			globalState.AgentState[agentDonation.agentId] = a
-		}
-	}
-
-	logging.Log(logging.Info, logging.LogField{
-		"Old HP Pool":           globalState.HpPool,
-		"HP Donated This Round": sum,
-		"New Hp Pool":           globalState.HpPool + sum,
-	}, "HP Pool Donation")
-
-	globalState.HpPool += sum
 }
